refactor(agents): implement shared ExponentialBackoffWithJitter mechanism
- Fixed unused variable compiler warnings (`resp` and `resp_json_am_fetch`) in `codex_runner.rs` by prefixing them with an underscore.
- Extracted `ExponentialBackoffWithJitter` and `RetryStrategy` from `output_parser.rs` into `types.rs` to implement a shared SOTA backoff mechanism globally.
- Updated `tool_executor_engine.rs` to replace its custom ad-hoc jitter with the shared `ExponentialBackoffWithJitter`.
- Cleanly deleted stranded test logic from `output_parser.rs` to prevent duplicate test issues in production files.
